drm: Boost DDR bus upon running an atomic ioctl
In order to reduce jank, request a DDR bus boost whenever an atomic
ioctl is sent, which is what our diplay HAL uses to commit frames to the
display. The boost should be sufficient enough to render the display's
maximum FPS without any dropped frames when there is no significant
external source of load.

[ghostrider-reborn]:
- Boost if only within 3.2s input timeout
- This prevents unnecessary boosting such as during video playback
- More details: https://github.com/caf-ginkgo/kernel_msm-4.14/commit/4ccf7265c0273f01c7fda6ba332ea78209bf9194

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Change-Id: Iaf2345c63554bfeb447a53168f1ff2487543c662
Signed-off-by: Excalibur-99 <txexcalibur99@gmail.com>